Tensile Plots from DIC
I have plotted the new stress curves with the corrected stress values.
There was not much shift between the stress and the strain in terms of
time so the ultimate stress strain curve looks better. However i am
unable to calculate the min/max point to correct for the little shift
that there is between the 2 curves